Extract HDR histogram implementation into a shared package
