+++ internal/ds/histogram/doc.go
@@ -1,0 +1,2 @@
+// Package histogram provides histogram implementations that are used to track the distribution of metrics.
+package histogram
